requirements.txt
langchain==0.1.12
langchain_google_genai==0.0.11
pandas==1.5.3
Pillow==10.2.0
plotly==5.20.0
protobuf==5.26.0
python-dotenv==1.0.1
streamlit==1.29.0
